Move helpers to utils and remove .d from type definition files
