+++ model/src/conditions/types.ts
