+++ model/test/list-item-references-to-value-objects.test.ts
@@ -3,7 +3,7 @@ import * as Lab from "@hapi/lab";
 import sinon from "sinon";
 import ListItemReferencesToValueObjects from "../src/migration/list-item-references-to-value-objects";
 
-import { clone } from "../src/helpers";
+import { clone } from "../src/utils/helpers";
 
 const { expect } = Code;
 const lab = Lab.script();
